requirements.txt
Django==1.9
argparse==1.2.1
dj-database-url==0.4.1
django-configurations==1.0
psycopg2==2.6.1
wsgiref==0.1.2
